Extract busy cursor-related stuff into its own header
It didn't make sense to include wx/utils.h from wx/cursor.h, but it
somewhat does make sense to declare busy cursor-related functions and
wxBusyCursor itself in this header, so extract these declarations into
their own header which is included from both wx/cursor.h and wx/utils.h
for compatibility, but can be also included directly.

Document that wx/cursor.h doesn't pull in wx/utils.h any longer as it's
an incompatible change (but hopefully one not resulting in too much
disruption and so worth it).
