+++ Makefile.in
@@ -3751,6 +3751,7 @@ COND_USE_GUI_1_ALL_GUI_HEADERS =  \
 	wx/artprov.h \
 	wx/bitmap.h \
 	wx/bookctrl.h \
+	wx/busycursor.h \
 	wx/busyinfo.h \
 	wx/generic/busyinfo.h \
 	wx/caret.h \
