+++ docs/changes.txt
@@ -231,6 +231,10 @@ Changes in behaviour which may result in build errors
   deriving from wxSizer, please change the function in the derived class to
   take wxWindowBase pointer too in this case.
 
+- wx/cursor.h doesn't include wx/utils.h any longer, please include it
+  explicitly if you need to use any functions defined there other than
+  cursor-related ones.
+
 
 3.3.0: (released 2022-??-??)
 ----------------------------
